Split comfyUIBackendService
