+++ WebUI/electron/main.ts
@@ -5,8 +5,18 @@ import { isAdmin } from './utils.ts'
 if (isAdmin()) {
   const lib = koffi.load('user32.dll')
   const MB_ICONINFORMATION = 0x40
-  const MessageBoxW = lib.func('__stdcall', 'MessageBoxW', 'int', ['void *', 'str16', 'str16', 'uint'])
-  MessageBoxW(null, 'For security reasons, AI Playground cannot be executed with administrative permissions. Please restart AI Playground from a Windows account without Administrator rights.', 'AI Playground', MB_ICONINFORMATION)
+  const MessageBoxW = lib.func('__stdcall', 'MessageBoxW', 'int', [
+    'void *',
+    'str16',
+    'str16',
+    'uint',
+  ])
+  MessageBoxW(
+    null,
+    'For security reasons, AI Playground cannot be executed with administrative permissions. Please restart AI Playground from a Windows account without Administrator rights.',
+    'AI Playground',
+    MB_ICONINFORMATION,
+  )
   process.exit(0)
 }
 
@@ -16,14 +26,19 @@ import fs from 'fs'
 import sudo from 'sudo-prompt'
 import { PathsManager } from './pathsManager'
 import { appLoggerInstance } from './logging/logger.ts'
-import { aiplaygroundApiServiceRegistry, ApiServiceRegistryImpl } from './subprocesses/apiServiceRegistry'
-import { loadSettings, getSettings } from './settings.ts'
+import {
+  aiplaygroundApiServiceRegistry,
+  ApiServiceRegistryImpl,
+} from './subprocesses/apiServiceRegistry'
+import { loadSettings, getSettings, type LocalSettings } from './settings.ts'
 import { createWindow, setupDisplayMetricsListener } from './window.ts'
 import { spawnLangchainUtilityProcess } from './langchain.ts'
 import { setupCoreIpcHandlers } from './ipcHandlers.ts'
 import { setupServiceIpcHandlers } from './ipcServiceHandlers.ts'
 import { setupComfyIpcHandlers } from './ipcComfyHandlers.ts'
 import { externalResourcesDir, getMediaDir, needAdminPermission } from './utils.ts'
+
+export type { LocalSettings }
 
 process.env.DIST = path.join(__dirname, '../')
 process.env.VITE_PUBLIC = path.join(__dirname, app.isPackaged ? '../..' : '../../../public')
@@ -37,17 +52,27 @@ let serviceRegistry: ApiServiceRegistryImpl | null = null
 const mediaDir = getMediaDir()
 fs.mkdirSync(mediaDir, { recursive: true })
 
-protocol.registerSchemesAsPrivileged([{
-  scheme: 'aipg-media',
-  privileges: { secure: true, supportFetchAPI: true, standard: true, bypassCSP: true, stream: true },
-}])
+protocol.registerSchemesAsPrivileged([
+  {
+    scheme: 'aipg-media',
+    privileges: {
+      secure: true,
+      supportFetchAPI: true,
+      standard: true,
+      bypassCSP: true,
+      stream: true,
+    },
+  },
+])
 
 app.on('quit', async () => {
   if (singleInstanceLock) app.releaseSingleInstanceLock()
 })
 
 app.on('window-all-closed', async () => {
-  try { await serviceRegistry?.stopAllServices() } catch {}
+  try {
+    await serviceRegistry?.stopAllServices()
+  } catch {}
   if (process.platform !== 'darwin') {
     app.quit()
     win = null
@@ -56,7 +81,9 @@ app.on('window-all-closed', async () => {
 
 app.on('activate', () => {
   if (BrowserWindow.getAllWindows().length === 0) {
-    createWindow(getSettings()).then((w) => { win = w })
+    createWindow(getSettings()).then((w) => {
+      win = w
+    })
   }
 })
 
@@ -78,8 +105,19 @@ function initEventHandlers() {
   )
 
   setupDisplayMetricsListener(win)
-  setupCoreIpcHandlers(pathsManager, getSettings(), mediaDir, externalRes, () => win, () => serviceRegistry?.getService('comfyui-backend')?.baseUrl)
-  setupServiceIpcHandlers(() => serviceRegistry, () => win, getSettings())
+  setupCoreIpcHandlers(
+    pathsManager,
+    getSettings(),
+    mediaDir,
+    externalRes,
+    () => win,
+    () => serviceRegistry?.getService('comfyui-backend')?.baseUrl,
+  )
+  setupServiceIpcHandlers(
+    () => serviceRegistry,
+    () => win,
+    getSettings(),
+  )
   setupComfyIpcHandlers(() => serviceRegistry, externalRes)
 }
 
@@ -87,13 +125,18 @@ app.whenReady().then(async () => {
   if (await needAdminPermission(externalRes)) {
     if (singleInstanceLock) app.releaseSingleInstanceLock()
     const message = `start "" "${process.argv.join(' ').trim()}`
-    sudo.exec(message, () => { app.exit(0) })
+    sudo.exec(message, () => {
+      app.exit(0)
+    })
     return
   }
 
   if (!singleInstanceLock) {
     dialog.showMessageBoxSync({
-      message: app.getLocale() == 'zh-CN' ? '本程序仅允许单实例运行，确认后本次运行将自动结束' : 'This program only allows a single instance to run, and the run will automatically end after confirmation',
+      message:
+        app.getLocale() == 'zh-CN'
+          ? '本程序仅允许单实例运行，确认后本次运行将自动结束'
+          : 'This program only allows a single instance to run, and the run will automatically end after confirmation',
       title: 'error',
       type: 'error',
     })
@@ -103,10 +146,15 @@ app.whenReady().then(async () => {
     initEventHandlers()
 
     protocol.handle('aipg-media', async (request) => {
-      const decodedUrl = decodeURIComponent(request.url.replace(/^aipg-media:\/\//i, '/'))
+      let decodedUrl = decodeURIComponent(request.url.replace(/^aipg-media:\/\//i, ''))
+      // Remove trailing slashes that might cause issues
+      decodedUrl = decodedUrl.replace(/\/+$/, '')
       const fullPath = path.join(mediaDir, decodedUrl)
-      const normalizedPath = path.normalize(fullPath.replace(/(\/|\\)$/, ''))
-      return await net.fetch(`file://${normalizedPath}`)
+      const normalizedPath = path.normalize(fullPath)
+      // Convert Windows path to file:// URL format
+      const fileUrl = normalizedPath.replace(/\\/g, '/')
+
+      return await net.fetch(`file:///${fileUrl}`)
     })
 
     win = await createWindow(settings)
@@ -119,4 +167,3 @@ app.whenReady().then(async () => {
     spawnLangchainUtilityProcess()
   }
 })
-
