+++ WebUI/electron/remoteUpdates.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import path from 'path'
 import { app } from 'electron'
 import { appLoggerInstance } from './logging/logger.ts'
-import { externalResourcesDir } from './util.ts'
+import { externalResourcesDir } from './utils.ts'
 import z from 'zod'
 import { ModelSchema } from '../src/types/shared.ts'
 import { LocalSettings } from './main.ts'
