--- WebUI/electron/util.ts
@@ -1,17 +1,0 @@
-import { app } from 'electron'
-import path from 'node:path'
-
-export const externalResourcesDir = () =>
-  path.resolve(app.isPackaged ? process.resourcesPath : path.join(__dirname, '../../external/'))
-
-export const getMediaDir = () => {
-  let mediaDir: string
-  if (process.env.USERPROFILE) {
-    mediaDir = path.join(process.env.USERPROFILE, 'Documents', 'AI-Playground', 'media')
-  } else if (process.env.HOME) {
-    mediaDir = path.join(process.env.HOME, 'AI-Playground', 'media')
-  } else {
-    mediaDir = path.join(externalResourcesDir(), 'service', 'static', 'sd_out')
-  }
-  return mediaDir
-}
