+++ WebUI/package.json
@@ -25,7 +25,6 @@
     "@langchain/community": "^1.1.15",
     "@langchain/core": "^1.1.24",
     "@langchain/openai": "^1.2.6",
-    "@langchain/textsplitters": "^1.0.5",
     "@radix-icons/vue": "^1.0.0",
     "@tailwindcss/vite": "^4.1.18",
     "@vee-validate/zod": "^4.15.1",
