+++ WebUI/src/assets/i18n/ja.json
@@ -324,7 +324,5 @@
   "BACKEND_SETTINGS_DESCRIPTION": "{backend}バックエンドの設定を構成します。変更を有効にするには再起動または再インストールが必要な場合があります。",
   "BACKEND_VERSION": "バージョン",
   "BACKEND_VERSION_DESCRIPTION": "バックエンドに使用するGit参照。これはバージョンタグまたはコミットハッシュです。",
-  "BACKEND_SAVE_CHANGES": "変更を保存",
-  "BACKEND_LOADING_TITLE": "起動中...",
-  "BACKEND_LOADING_MESSAGE": "準備ができ次第、リクエストが自動的に処理されます"
+  "BACKEND_SAVE_CHANGES": "変更を保存"
 }
